Added A-packages: Organised the classes into suitable java packages
