--- src/main/java/CommandType.java
@@ -1,12 +1,0 @@
-public enum CommandType {
-    LIST,
-    DONE,
-    DELETE,
-    TODO,
-    DEADLINE,
-    EVENT,
-    BYE,
-    UNKNOWN,
-    SHOW_DATE
-
-}
