+++ src/main/java/boblet/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package boblet.command;
+
+import boblet.util.Storage;
+import boblet.util.TaskList;
+import boblet.util.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
