+++ src/main/java/boblet/exception/BobletException.java
@@ -1,3 +1,4 @@
+package boblet.exception;
 public class BobletException extends Exception {
     public BobletException(String message) {
         super(message);
