+++ src/main/java/boblet/task/TaskType.java
@@ -1,3 +1,4 @@
+package boblet.task;
 public enum TaskType {
     TODO,
     DEADLINE,
