+++ src/main/java/boblet/task/Todo.java
@@ -1,3 +1,5 @@
+package boblet.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskType.TODO);
