+++ src/main/java/boblet/util/Ui.java
@@ -1,3 +1,4 @@
+package boblet.util;
 import java.util.Scanner;
 
 public class Ui {
